refactor: 불필요한 매개변수 제거
